pyproject.toml
[tool.poetry]
name = "saber-test"
version = "0.1.0"
description = ""
authors = ["Artur <ArthurZ43@yandex.ru>"]
readme = "README.md"
packages = [{include = "saber_test"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
dependency-injector = "^4.41.0"
pyyaml = "^6.0"
uvicorn = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
